fix: wire integration tests
