+++ packages/integration-tests/src/shared/templates.js
@@ -13,45 +13,8 @@ exports.app = function(cmpName) {
 
 exports.todoApp = function(cmpName) {
     return `
-        import { registerWireService, register as registerAdapter, ValueChangedEvent } from 'wire-service';
-        import { createElement, register } from 'lwc';
+        import { createElement } from 'lwc';
         import Cmp from 'integration/${cmpName}';
-        import { getTodo, getObservable } from 'todo';
-
-        registerWireService(register);
-
-        // Register the wire adapter for @wire(getTodo).
-        registerAdapter(getTodo, function getTodoWireAdapter(wiredEventTarget) {
-            var subscription;
-            var config;
-            wiredEventTarget.dispatchEvent(new ValueChangedEvent({ data: undefined, error: undefined }));
-            var observer = {
-                next: function(data) { wiredEventTarget.dispatchEvent(new ValueChangedEvent({ data: data, error: undefined })); },
-                error: function(error) { wiredEventTarget.dispatchEvent(new ValueChangedEvent({ data: undefined, error: error })); }
-            };
-            wiredEventTarget.addEventListener('connect', function() {
-                var observable = getObservable(config);
-                if (observable) {
-                    subscription = observable.subscribe(observer);
-                    return;
-                }
-            });
-            wiredEventTarget.addEventListener('disconnect', function() {
-                subscription.unsubscribe();
-            });
-            wiredEventTarget.addEventListener('config', function(newConfig) {
-                config = newConfig;
-                if (subscription) {
-                    subscription.unsubscribe();
-                    subscription = undefined;
-                }
-                var observable = getObservable(config);
-                if (observable) {
-                    subscription = observable.subscribe(observer);
-                    return;
-                }
-            });
-        });
 
         var element = createElement('integration-${cmpName}', { is: Cmp });
         document.body.appendChild(element);
